+++ changelog.md
@@ -3,6 +3,7 @@
 - iOS support!
 - Fix issues with the pause keybind not updating the button in the current song popup
 - Fix issues with the current song popup not closing if you closed the editor pause menu
+- Fix issues with songs in history not being preloaded when they should
 ## v2.1.14
 - Fix songs longer than one hour or take up more space than memory available on the system
 - Make reading metadata more consistent and less crashable when filtering invalid characters